WillowTree Software Test Engineer Test Project - Rudy Salas
Added an up-to-date version of the chromedriver to be used with my linux
machine.

Added validate*() methods to HomePage for the following test scenarios:
<ul>
<li>Verify that the "streak" counter is incrementing on correct selections.</li>
<li>Verify that a  multiple “streak” counter resets after getting an incorrect answer.</li>
<li>Verify that after 10 random selections the correct counters are being incremented for tries and correct counters.</li>
<li>Verify name and displayed photos change after selecting the correct answer.</li>
</ul>

Added calls to each validate*() method within the WebTest class.

Added "helper" methods which represent elements of the page which are
commonly used, and methods for retrieving the names of the people in the
gallery and the WebElement represented by a given name.